added new make target and removed build errors

+++ test/rm_test_util.h
@@ -367,7 +367,7 @@ void writeRIDsToDisk(vector<RID> &rids) {
 
     if (ridsFile.is_open()) {
         ridsFile.seekp(0, ios::beg);
-        for (int i = 0; i < rids.size(); i++) {
+        for (size_t i = 0; i < rids.size(); i++) {
             ridsFile.write(reinterpret_cast<const char *>(&rids[i].pageNum),
                            sizeof(unsigned));
             ridsFile.write(reinterpret_cast<const char *>(&rids[i].slotNum),
@@ -385,7 +385,7 @@ void writeSizesToDisk(vector<int> &sizes) {
 
     if (sizesFile.is_open()) {
         sizesFile.seekp(0, ios::beg);
-        for (int i = 0; i < sizes.size(); i++) {
+        for (size_t i = 0; i < sizes.size(); i++) {
             //cout << sizes[i] << endl;
             sizesFile.write(reinterpret_cast<const char *>(&sizes[i]),
                             sizeof(int));
